Created Form
Created form, added two text boxes, one combo box, and one button. User can type or paste text in the top (input) box. Upon clicking the Call API button, the text will be transferred to the bottom (output) box. The combo box selects whether the text in the input box will remain or be erased; the default value is to keep the text.